requirements.txt
# ShadowBench Enterprise - Production Dependencies
# Core Framework Requirements for Enterprise AI Security

# ============================================================================
# CORE FRAMEWORK DEPENDENCIES (REQUIRED)
# ============================================================================

# Python Core Libraries
pyyaml>=6.0.1
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.1.0
requests>=2.31.0
urllib3>=2.0.0
dataclasses-json>=0.6.0

# CLI & Configuration
click>=8.1.0
rich>=13.7.0
tabulate>=0.9.0
colorlog>=6.8.0
tqdm>=4.66.0

# Data Processing & Validation
jsonschema>=4.20.0
jsonlines>=4.0.0
python-dateutil>=2.8.0
pathlib>=1.0.1

# Cryptography & Security
cryptography>=41.0.0
hashlib2>=1.0.1
bcrypt>=4.1.0
pynacl>=1.5.0

# ============================================================================
# TESTING & QUALITY ASSURANCE (REQUIRED FOR PRODUCTION)
# ============================================================================

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
pytest-timeout>=2.2.0
pytest-mock>=3.12.0

# Code Quality Tools  
flake8>=6.1.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0
bandit>=1.7.0
safety>=2.3.0

# Performance Testing
locust>=2.18.0
memory-profiler>=0.61.0

# ============================================================================
# ADVERSARIAL AI & MACHINE LEARNING (CORE FEATURES)
# ============================================================================

# Scientific Computing
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# NLP & Text Processing
nltk>=3.8.1
spacy>=3.7.0
transformers>=4.36.0
sentence-transformers>=2.2.0
rouge-score>=0.1.2
evaluate>=0.4.0

# Computer Vision (OpenCV alternative for security)
opencv-python-headless>=4.8.0
Pillow>=10.1.0
imageio>=2.33.0
scikit-image>=0.22.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.11

# ============================================================================
# MODEL API INTEGRATIONS (ENTERPRISE FEATURES)
# ============================================================================

# Major API Providers
openai>=1.6.0
anthropic>=0.8.0
google-cloud-aiplatform>=1.38.0
cohere>=4.39.0
replicate>=0.22.0

# Local Model Inference
torch>=2.1.0
torchvision>=0.16.0
accelerate>=0.25.0
sentencepiece>=0.1.99
tiktoken>=0.5.0

# Optional: High-Performance Local Inference
# vllm>=0.2.7  # Uncomment for local LLM inference
# llama-cpp-python>=0.2.0  # Uncomment for CPU inference

# ============================================================================
# ENTERPRISE WEB & API FRAMEWORK (DASHBOARD)
# ============================================================================

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-restx>=1.3.0
gunicorn>=21.2.0
waitress>=2.1.0

# Database Connectors
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0
redis>=5.0.0

# Authentication & Security
flask-jwt-extended>=4.6.0
flask-login>=0.6.0
werkzeug>=3.0.0
itsdangerous>=2.1.0

# ============================================================================
# MONITORING & OBSERVABILITY (PRODUCTION ESSENTIAL)
# ============================================================================

# System Monitoring
psutil>=5.9.0
GPUtil>=1.4.0
py3nvml>=0.2.7

# Application Performance
prometheus-client>=0.19.0
statsd>=4.0.0

# Logging & Tracing
structlog>=23.2.0
python-json-logger>=2.0.0
opentelemetry-api>=1.21.0

# Optional: Advanced Monitoring
# wandb>=0.16.0  # Uncomment for experiment tracking
# sentry-sdk>=1.39.0  # Uncomment for error tracking

# ============================================================================
# CONTAINERIZATION & DEPLOYMENT (DEVOPS)
# ============================================================================

# Docker Integration
docker>=6.1.0
docker-compose>=1.29.0

# Cloud SDK (Optional)
# boto3>=1.34.0  # AWS
# azure-storage-blob>=12.19.0  # Azure
# google-cloud-storage>=2.13.0  # GCP

# ============================================================================
# DEVELOPMENT & JUPYTER INTEGRATION (OPTIONAL)
# ============================================================================

# Interactive Development
ipython>=8.18.0
jupyter>=1.0.0
notebook>=7.0.0
jupyterlab>=4.0.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
myst-parser>=2.0.0

# ============================================================================
# VERSION CONSTRAINTS FOR SECURITY & COMPATIBILITY
# ============================================================================

# Pin critical security packages
urllib3>=2.0.7  # Security fix
certifi>=2023.11.17  # Latest CA bundle
setuptools>=69.0.0  # Security updates

# Python Version Compatibility
# Requires Python 3.10+ for best performance and security
